# magma-oas/src/main/resources/openapi-gestion.yaml
openapi: 3.1.0
info:
  title: Magma Gestion API
  description: API de gestion des métadonnées statistiques
  contact:
    name: dr59-sndil
    url: https://www.insee.fr/fr/accueil
    email: dr59-sinl-equipe-maintenance-rmes@insee.fr
  version: 1.4.11
...
security:
  - bearerAuth: []
paths:
...
  /structures:
    get:
      tags:
        - Structures/Composants
      summary: Get all structures
      operationId: getAllStructures
      parameters:
        - name: dateMiseAJour
          in: query
          description: 'Date of last update. Example: 2023-01-31'
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AllStructure'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /structure/{id}:
    get:
      tags:
        - Structures/Composants
      summary: Get a structure
      operationId: getStructure
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: dateMiseAJour
          in: query
          description: If true, returns only the update date
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructureById'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /structure/{id}/sliceKeys:
    get:
      tags:
        - Structures/Composants
      summary: Get slice keys
      operationId: getSliceKeys
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StructureSliceKeys'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /composants:
    get:
      tags:
        - Structures/Composants
      summary: Get all components
      operationId: getAllComponents
      parameters:
        - name: dateMiseAJour
          in: query
          description: 'Date of last update. Example: 2023-01-31'
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AllComponent'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /composant/{id}:
    get:
      tags:
        - Structures/Composants
      summary: Get a component
      operationId: getComponentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: dateMiseAJour
          in: query
          description: If true, returns only the update date
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentById'
        '404':
          description: Not found
        '500':
          description: Internal server error

  # ============================================
  # ORGANISATIONS
  # ============================================
...
  /operations/series:
    get:
      tags:
        - Series-Operations
      summary: Get all series
      operationId: getAllSeries
      parameters:
        - name: survey
          in: query
          description: Filter for survey only
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SerieById'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /operations/serie/{id}:
    get:
      tags:
        - Series-Operations
      summary: Get one serie
      operationId: getSerieById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SerieById'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /operations/serie/{id}/operations:
    get:
      tags:
        - Series-Operations
      summary: Get operations by serie
      operationId: getOperationsBySerie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperationBySerieId'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /operations/operation/{id}:
    get:
      tags:
        - Series-Operations
      summary: Get operation by code
      operationId: getOperationByCode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationById'
        '404':
          description: Not found
        '500':
          description: Internal server error

  # ============================================
  # CODES LISTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Label:
      type: object
      properties:
        langue:
          type: string
        contenu:
          type: string
...
    AllStructure:
      type: object
      properties:
        dateMiseAJour:
          type: string
        statutValidation:
          type: string
        id:
          type: string
    StructureById:
      type: object
      properties:
        dateMiseAJour:
          type: string
        dateCreation:
          type: string
        notation:
          type: string
        statutValidation:
          type: string
        mesures:
          type: array
          items:
            $ref: '#/components/schemas/Mesure'
        id:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
        version:
          type: string
        attributs:
          type: array
          items:
            $ref: '#/components/schemas/Attribut'
        dimensions:
          type: array
          items:
            $ref: '#/components/schemas/Dimension'
    StructureSliceKeys:
      type: object
      properties:
        composants:
          type: array
          items:
            type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        type:
          type: string
    Mesure:
      type: object
      properties:
        ordre:
          type: string
        notation:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
        representation:
          type: string
    Attribut:
      type: object
      properties:
        ordre:
          type: string
        notation:
          type: string
        listCode:
          $ref: '#/components/schemas/ListCodeRef'
        attachement:
          type: string
        obligatoire:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
        representation:
          type: string
    Dimension:
      type: object
      properties:
        ordre:
          type: string
        notation:
          type: string
        listCode:
          $ref: '#/components/schemas/ListCodeRef'
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
        representation:
          type: string
    ListCodeRef:
      type: object
      properties:
        id:
          type: string
        uri:
          type: string
    AllComponent:
      type: object
      properties:
        dateMiseAJour:
          type: string
        statutValidation:
          type: string
        id:
          type: string
    ComponentById:
      type: object
      properties:
        dateMiseAJour:
          type: string
        dateCreation:
          type: string
        concept:
          $ref: '#/components/schemas/ConceptRef'
        statutValidation:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        type:
          type: string
        uri:
          type: string
        representation:
          type: string
        version:
          type: string
        notation:
          type: string
        id:
          type: string
        listeCode:
          $ref: '#/components/schemas/ListeCodeWithCodes'
    ConceptRef:
      type: object
      properties:
        id:
          type: string
        uri:
          type: string
    ListeCodeWithCodes:
      type: object
      properties:
        codes:
          type: array
          items:
            $ref: '#/components/schemas/Code'
        id:
          type: string
        uri:
          type: string

    # ============================================
    # ORGANISATIONS SCHEMAS
    # ============================================
...
    SerieById:
      type: object
      properties:
        seriesId:
          type: string
        uri:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        altLabel:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        resume:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        noteHistorique:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        type:
          $ref: '#/components/schemas/TypeRef'
        frequenceCollecte:
          $ref: '#/components/schemas/TypeRef'
        rapportQualite:
          $ref: '#/components/schemas/RapportQualite'
        famille:
          $ref: '#/components/schemas/Famille'
        seriesPrecedentes:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        seriesSuivantes:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        seriesLiees:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        operations:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        indicateurs:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        dateCreation:
          type: string
          format: date
        dateMiseAJour:
          type: string
          format: date
        proprietaires:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        organismesResponsables:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        partenaires:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        servicesCollecteurs:
          type: array
          items:
            $ref: '#/components/schemas/TypeRef'
        statutValidation:
          type: string
        series:
          type: string
        frequence:
          $ref: '#/components/schemas/Frequence'
        nbOperation:
          type: string
    RapportQualite:
      type: object
      properties:
        id:
          type: string
        uri:
          type: string
    TypeRef:
      type: object
      properties:
        id:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
    Frequence:
      type: object
      properties:
        id:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
    Famille:
      type: object
      properties:
        id:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
    OperationBySerieId:
      type: object
      properties:
        altLabel:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
        serie:
          $ref: '#/components/schemas/SerieRef'
        id:
          type: string
        millesime:
          type: string
    SerieRef:
      type: object
      properties:
        id:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string
        frequence:
          $ref: '#/components/schemas/Frequence'
    OperationById:
      type: object
      properties:
        id:
          type: string
        uri:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        altLabel:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        millesime:
          type: string
        serie:
          $ref: '#/components/schemas/SerieRef'
        rapportQualite:
          $ref: '#/components/schemas/RapportQualite'
        dateCreation:
          type: string
          format: date
        dateMiseAJour:
          type: string
          format: date
        statutValidation:
          type: string

    # ============================================
    # CODE LISTS SCHEMAS
    # ============================================
...
    Code:
      type: object
      properties:
        code:
          type: string
        label:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        uri:
          type: string

    # ============================================
    # DATASETS SCHEMAS
    # ============================================
...